finalized all data creation
Got the generate tfrecords file making correct output. It now makes the normal data along with black and white data with 50, 100, 150, 200, and 250 thresholds so we can see which has the higher accuracy.

Showing samples does not completely work, but we do not need it now since generating outputs files. Still should see if this can get working in case we run into issues down the line with reading files.